feat: Add idea for jwt check in lambda functions
